requirements.txt
numpy
opencv-python
PyYAML
matplotlib